requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pixelblaze-client==1.1.3
py_mini_racer==0.6.0
python-osc==1.8.3
pytz==2024.1
requests==2.32.3
urllib3==2.2.1
websocket-client==1.8.0
